Separated searcher into its own component
