+++ TODO.txt
@@ -1,0 +1,8 @@
+Things that need fixing!
+
+- Input text doesn't empty after submit
+
+- When a wrong value is entered, it changes the name but it doesn't update the temperature,
+so it is unclear if it has found it or not
+   - Erase old info
+   - Configure an error
